Snoozing status now stored in DV List and updates via monitor thread
Monitor thread udpates all assigned decvice to snooze status change and
updates the cached DV list when an alteration is detected. As a side
effect this has also elminated the need for hooks as the monitor updates
the DVList when a DV is found that it has no record of.