Device GCM key now automatically updates cancel thread issue handled
Logging in devices now updates the stored associated GCM key to that
transmitted with the phone. The new key is paired with the existing
android id in DB and all associated threads are restarted to include the
new key for the device in order to ensure coverage. Bug found where
removal of thread still left instance of thead messaging has been
resolved and been handled on both ends.